requirements.txt
Flask==3.0.0
Pillow==10.1.0
numpy==1.26.0
torch==2.1.0
torchvision==0.16.0
scipy==1.11.3
matplotlib==3.8.0